requirements.txt
tensorflow==2.13.0
tensorflow-addons==0.21.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.76
Pillow==10.0.0
albumentations==1.3.1
imgaug==0.4.0
tqdm==4.65.0
jupyter==1.0.0
ipykernel==6.25.0
plotly==5.15.0
wandb==0.15.8
tensorboard==2.13.0
flask==2.3.2
fastapi==0.101.1
uvicorn==0.23.2
python-multipart==0.0.6
aiofiles==23.1.0
python-dotenv==1.0.0
pydantic==2.1.1
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.3.3
